refactor(server): extract /api/geo-strategist/* into a router module
Seventh route-group extraction; first of the harder mixed-auth tier.
3 handlers (briefs list, briefs/:id read, analyze) -> src/server/routes/
geo-strategist.js, mounted `app.use('/api/geo-strategist', geoStrategistRouter)`.

Auth: MIXED — GET /briefs/:id is public, briefs + analyze are requireAuth
-> per-route (no mount-level), requireAuth imported and kept on the 2.

Deps all already extracted: pool, anthropic, extractJSON (llm-json),
normalizeGeoData (geo.js), requireAuth (auth) + randomUUID (crypto).
No local-helper blockers. Handler bodies verbatim.

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean (caught randomUUID; added)
- route-inventory guard -> 213 byte-identical (3 paths)
- vitest -> 93/93

Rollback: revert; verbatim move.
